server/requirements.txt
Flask==2.3.2
flask_jwt_extended==4.7.1
psycopg2-binary==2.9.8